Refactored mistakes before immediate launch

+++ controllers/settings/settings.ts
@@ -2,6 +2,16 @@ import { Response, Request, NextFunction } from "express";
 import { Settings } from "../../models/settings";
 import { User } from "../../models/user";
 import { sendSuccess } from "../../utils/sendSuccess";
+import { sequelize } from "../../database/pool";
+import { Token } from "../../models/token";
+import { Account } from "../../models/better-auth";
+import { Pointer } from "../../models/pointer";
+import { File } from "../../models/file";
+import { Folder } from "../../models/folder";
+import { Activity } from "../../models/activity";
+import { UserJob } from "../../models/user-jobs";
+import { Subscription } from "../../models/subscription";
+import { Transaction } from "../../models/transaction";
 
 export async function FetchSettingsData(
   req: Request,
@@ -76,16 +86,35 @@ export async function deleteAccount(
   res: Response,
   next: NextFunction,
 ) {
+  const t = await sequelize.transaction();
   try {
     const decoded = req.user;
     const userId = decoded?.sub;
 
-    await Settings.destroy({
-      where: { userId },
-    });
+    await Transaction.destroy({ where: { userId }, transaction: t });
+    await Subscription.destroy({ where: { userId }, transaction: t });
+    await UserJob.destroy({ where: { userId }, transaction: t });
+    await Activity.destroy({ where: { userId }, transaction: t });
+
+    const pointers = await Pointer.findAll({ where: { userId }, transaction: t });
+    const pointerIds = pointers.map((p) => p.id);
+
+    if (pointerIds.length > 0) {
+      await File.destroy({ where: { id: pointerIds }, transaction: t });
+      await Folder.destroy({ where: { id: pointerIds }, transaction: t });
+      await Pointer.destroy({ where: { userId }, transaction: t });
+    }
+
+    await Account.destroy({ where: { userId }, transaction: t });
+    await Token.destroy({ where: { userId }, transaction: t });
+    await Settings.destroy({ where: { userId }, transaction: t });
+    await User.destroy({ where: { id: userId }, transaction: t });
+
+    await t.commit();
 
     sendSuccess(res, undefined, undefined, "ACCOUNT_DELETED");
   } catch (error) {
+    await t.rollback();
     next(error);
   }
 }
